# qa-shop/api/openapi.yaml
openapi: 3.0.3
...
info:
  title: QA Shop Practice API
  version: "1.0.0"
  description: |
    Test otomasyonu pratiği için gerçek bir e-ticaret API'si. Mock değil:
    veri kalıcı, işlemler transaction içinde, stok gerçekten düşüyor.

    **Sandbox anahtarı göndermezsen** demo verisine salt okunur bağlanırsın —
    `GET` uçları çalışır, yazma denemeleri `401` döner. İlk denemeyi
    kolaylaştırmak için böyle.

    **Kendi alanını açmak için:** `POST /sandbox` → dönen `apiKey` değerini
    `X-Sandbox-Key` başlığında gönder. Alan 7 gün sonra otomatik silinir.

    **Her test koşumundan önce** `POST /sandbox/reset` çağır — temiz durum
    otomasyonun en temel disiplinidir ve burada bir uç olarak hazır.

    Demo hesabı: `demo@qashop.test` / `Password123!`
  contact:
    name: LearnQA.dev
    url: https://learnqa.dev
servers:
  - url: http://localhost:4000
    description: Lokal Docker (docker compose up)
...
security:
  - SandboxKey: []
paths:
...
  /api/v1/sandbox:
    post:
      tags: [Sandbox]
      summary: Yeni izole veri alanı aç
      description: |
        Kayıt İSTEMEZ. Şablonun birebir kopyasını alırsın: 120 ürün, 360
        varyant, 41 kullanıcı, 150 siparişlik geçmiş. Yani SQL pratiği ilk
        saniyeden itibaren anlamlı hacim bulur.
      security: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                label:
                  type: string
                  maxLength: 80
                  example: benim-pratik-alanim
      responses:
        "201":
          description: Alan açıldı
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SandboxCreated" }
  /api/v1/sandbox/state:
    get:
      tags: [Sandbox]
      summary: Satır sayıları
      description: |
        Test öncesi/sonrası kıyas için. "Testim gerçekten bir şey değiştirdi
        mi?" sorusunun en hızlı cevabı.
      responses:
        "200":
          description: Durum özeti
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SandboxState" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "410": { $ref: "#/components/responses/SandboxExpired" }
  /api/v1/sandbox/reset:
    post:
      tags: [Sandbox]
      summary: Tohum veriye dön
      description: |
        Tüm kullanıcı verisi silinir ve şablon yeniden kopyalanır. Açık
        oturumlar da silinir — sıfırlamadan sonra tekrar `login` gerekir.

        Test paketinin `beforeAll` adımına konması beklenir.
      responses:
        "200":
          description: Sıfırlandı
          content:
            application/json:
              schema:
                type: object
                properties:
                  sandboxId: { type: string, format: uuid }
                  resetAt: { type: string, format: date-time }
                  durationMs: { type: integer, example: 340 }
                  message: { type: string }
        "401":
          description: Demo (salt okunur) modda sıfırlama yapılamaz
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/v1/sandbox/logs:
    get:
      tags: [Sandbox]
      summary: Denetim kaydı (log analizi pratiği)
      description: |
        Bir hata aldığında cevabın `correlationId` değerini buraya ver;
        isteğin tüm zincirini görürsün. Kök neden analizinin API tarafı budur.
      parameters:
        - name: level
          in: query
          schema: { type: string, enum: [INFO, WARN, ERROR] }
        - name: action
          in: query
          schema: { type: string, example: order.create }
        - name: correlationId
          in: query
          schema: { type: string, example: req-8f3c1e2a }
        - name: limit
          in: query
          schema: { type: integer, default: 100, maximum: 500 }
      responses:
        "200":
          description: Log satırları (en yeni önce)
          content:
            application/json:
              schema:
                type: object
                properties:
                  total: { type: integer }
                  limit: { type: integer }
                  logs:
                    type: array
                    items: { $ref: "#/components/schemas/AuditLog" }
        "400":
          description: Geçersiz `level` değeri
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }

  # ─────────────────────────────────── Kimlik ────────────────────────────────
...
  /api/v1/sandbox/bugs:
    get:
      tags: [Kusur]
      summary: Açılabilecek kusurların kataloğu
      description: |
        Her anahtar sistemin gerçek bir yerinde gerçek bir kusur açar.
        `catchableBy` alanı, o kusuru hangi kontrolün yakalaması GEREKTİĞİNİ
        söyler.

        Pratiğin kendisi tam olarak şudur: anahtarı aç, testini koş, testinin
        kırmızıya dönüp dönmediğine bak. Dönmüyorsa test bir şeye bakmıyordur.
      responses:
        "200":
          description: Katalog ve açık olanlar
          content:
            application/json:
              schema:
                type: object
                properties:
                  sandboxId: { type: string, format: uuid }
                  active:
                    type: array
                    items: { type: string }
                  available:
                    type: array
                    items: { $ref: "#/components/schemas/BugFlag" }
    patch:
      tags: [Kusur]
      summary: Kusur aç / kapat
      description: |
        Kısmi güncelleme: gönderilmeyen anahtarlar olduğu gibi kalır.

        Bilinmeyen anahtar SESSİZCE yok sayılmaz, `UNKNOWN_BUG_FLAG` ile 422
        döner. Yok sayılsaydı yazım hatası yapan kullanıcı kusuru açtığını sanır,
        testi yeşil kalır ve bundan "testim çalışıyor" sonucunu çıkarırdı — tam
        olarak önlemeye çalıştığımız yanılgı.

        `POST /sandbox/reset` tüm anahtarları kapatır.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: { type: boolean }
              example: { oversell: true, discount_twice: false }
      responses:
        "200":
          description: Güncel anahtar durumu
          content:
            application/json:
              schema:
                type: object
                properties:
                  sandboxId: { type: string, format: uuid }
                  active:
                    type: array
                    items: { type: string }
                  flags: { type: object, additionalProperties: { type: boolean } }
        "401": { $ref: "#/components/responses/ReadOnlySandbox" }
        "422":
          description: "`UNKNOWN_BUG_FLAG`, `INVALID_FLAG_VALUE` veya `EMPTY_BODY`"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }

# ============================================================================
components:
  securitySchemes:
    SandboxKey:
      type: apiKey
      in: header
      name: X-Sandbox-Key
      description: |
        Hangi veri alanına bağlanacağını belirler. `POST /api/v1/sandbox` ile
        alınır. Gönderilmezse demo verisine SALT OKUNUR bağlanılır.
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        `POST /api/v1/auth/login` ile alınır. Gerçek JWT formatındadır —
        jwt.io'ya yapıştırıp içindeki `sub`, `sandbox`, `jti`, `exp`
        alanlarını görebilirsin.
  responses:
    Unauthorized:
      description: Token yok, geçersiz, süresi dolmuş ya da oturum iptal edilmiş
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    ReadOnlySandbox:
      description: Demo verisi salt okunurdur — önce `POST /api/v1/sandbox`
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    SandboxExpired:
      description: |
        Sandbox süresi doldu. `410 Gone`, `404`ten daha bilgilendiricidir:
        kaynak VARDI, süresi geçti.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      description: |
        Hata gövdesi HER uçta aynı şekildedir. Her endpoint kendi hata şeklini
        döndürseydi test kodu her uç için ayrı yazılır ve kırılgan olurdu.
      properties:
        error:
          type: object
          properties:
            code: { type: string, example: OUT_OF_STOCK }
            message: { type: string, example: Stokta yalnızca 2 adet var }
            details:
              type: object
              nullable: true
              example: { requested: 5, available: 2 }
        correlationId:
          type: string
          example: req-8f3c1e2a-...
          description: Cevabın `X-Correlation-Id` başlığında da döner.
...
    SandboxCreated:
      type: object
      properties:
        sandboxId: { type: string, format: uuid }
        apiKey: { type: string, example: qas_3f9a2b... }
        label: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
        expiresAt: { type: string, format: date-time }
        howToUse: { type: string }
        demoUser:
          type: object
          properties:
            email: { type: string, example: demo@qashop.test }
            password: { type: string, example: Password123! }
    SandboxState:
      type: object
      properties:
        sandboxId: { type: string, format: uuid }
        mode: { type: string, enum: [demo-readonly, private] }
        lastResetAt: { type: string, format: date-time, nullable: true }
        counts:
          type: object
          properties:
            products: { type: integer, example: 120 }
            variants: { type: integer, example: 360 }
            users: { type: integer, example: 41 }
            carts: { type: integer }
            cart_items: { type: integer }
            orders: { type: integer, example: 150 }
            order_items: { type: integer }
            reviews: { type: integer, example: 200 }
            audit_log: { type: integer }
    AuditLog:
      type: object
      properties:
        id: { type: integer }
        at: { type: string, format: date-time }
        level: { type: string, enum: [INFO, WARN, ERROR] }
        actor: { type: string, nullable: true }
        action: { type: string, example: order.create }
        entity: { type: string, nullable: true }
        entity_id: { type: string, nullable: true }
        correlation_id: { type: string, nullable: true }
        detail: { type: object, nullable: true }
...
    BugFlag:
      type: object
      description: |
        Kontrollü bir kusur. `breaks` neyi bozduğunu, `catchableBy` hangi
        kontrolün onu yakalaması gerektiğini söyler.
      properties:
        key: { type: string, example: oversell }
        enabled: { type: boolean }
        title:
          type: object
          properties:
            tr: { type: string }
            en: { type: string }
        breaks:
          type: object
          properties:
            tr: { type: string }
            en: { type: string }
        catchableBy:
          type: object
          properties:
            tr: { type: string }
            en: { type: string }
        reconciled:
          type: boolean
          description: |
            Satır toplamları başlıktaki ara toplama eşit mi. `false` ise
            veride mutabakat hatası var demektir.